inventarios, configurar producto

+++ src/app/components/medida/medida.component.ts
@@ -7,6 +7,9 @@ import { Medida } from 'src/app/interfaces/medida/medida';
 // Services
 import { InventarioService } from 'src/app/services/inventario.service';
 
+// SweetAlert
+import sweetalert from 'sweetalert';
+
 @Component({
   selector: 'app-medida',
   templateUrl: './medida.component.html',
@@ -17,7 +20,7 @@ export class MedidaComponent implements OnInit {
   myForm: FormGroup;
   @ViewChild('testButton', { static: false }) testButton: ElementRef;
 
-  constructor(private inventarioService: InventarioService) { 
+  constructor(private inventarioService: InventarioService) {
     this.myForm = new FormGroup({
       _medida: new FormControl('', [Validators.required])
     })
@@ -29,21 +32,21 @@ export class MedidaComponent implements OnInit {
 
   medidas: Medida[] = [];
 
-  consultarMedidas(){
+  consultarMedidas() {
     this.inventarioService.consultarMedidas(
       localStorage.getItem('miCuenta.getToken')
     )
-    .then(
-      ok => {
-        this.medidas = [];
-        this.medidas = ok['respuesta'];
-      }
-    )
-    .catch(
-      error => {
-        console.log(error);
-      }
-    )
+      .then(
+        ok => {
+          this.medidas = [];
+          this.medidas = ok['respuesta'];
+        }
+      )
+      .catch(
+        error => {
+          console.log(error);
+        }
+      )
   }
 
   validarFormulario() {
@@ -90,37 +93,48 @@ export class MedidaComponent implements OnInit {
       this.myForm.get('_medida').value,
       localStorage.getItem('miCuenta.putToken')
     )
-    .then(
-      ok => {
-        this.myForm.reset();
-        this.testButton.nativeElement.value = 'ingresar';
-        this.consultarMedidas();
-      }
-    )
-    .catch(
-      error => {
-        console.log(error);
-      }
-    )
+      .then(
+        ok => {
+          this.myForm.reset();
+          this.testButton.nativeElement.value = 'ingresar';
+          this.consultarMedidas();
+        }
+      )
+      .catch(
+        error => {
+          console.log(error);
+        }
+      )
   }
 
   eliminarMedida(idMedida) {
-    this.inventarioService.eliminarMedida(
-      idMedida,
-      localStorage.getItem('miCuenta.deleteToken')
-    )
-    .then(
-      ok => {
-        this.consultarMedidas();
-      }
-    )
-    .catch(
-      error => {
-        console.log(error);
-      }
-    )
+    sweetalert({
+      title: "Advertencia",
+      text: "¿Está seguro que desea eliminar?",
+      icon: "warning",
+      buttons: ['Cancelar', 'Ok'],
+      dangerMode: true
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          this.inventarioService.eliminarMedida(
+            idMedida,
+            localStorage.getItem('miCuenta.deleteToken')
+          )
+            .then(
+              ok => {
+                this.consultarMedidas();
+              }
+            )
+            .catch(
+              error => {
+                console.log(error);
+              }
+            )
+        }
+      });
   }
-  
+
   get _medida() {
     return this.myForm.get('_medida');
   }
